+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All changes to this project will be documented in this file.
 
+## [1.3.0] - 2026-04-06
+
+### Added
+
+- Added instance mininum and maximum level to the instance grid
+- Added instance sorting (level and name) to the instance grid
+
+### Fixed
+
+- Fixed difficulty color of instance always being orange if the instance had the same min and max level
+
 ## [1.2.5] - 2026-04-03
 
 ### Added
@@ -20,7 +31,7 @@ All changes to this project will be documented in this file.
 
 ### Added
 
-- Added Egg Collection as a required quest
+- Added Egg Collection as a pre-required quest
 
 ## [1.2.2] - 2026-04-02